Fix failing import of ExampleQWidget
An  import remained in the  file, after I had changed the name of the example widget from the cookiecutter to . This caused problems when opening the plugin for the first time. Noticable when opening an h5 file right after opening napari or when clicking on 'Melt Pool Tracker' in the plugin menu before opening an h5 file.