Add steps for surface annotation
This commit implements the three steps necessary to go from the window resliced image to surface feature annotation of the edges of the melt pool and the depression zone.The steps are as follows:
1. Clear the background with Otsu thresholding and apply a median filter to reduce noise.
2. Calculate the radial gradient centered at the laser position.
3. Extract the surface and add shapes layers necessary for the annotation.

In addition to this, I implemented a crude version of the only showing the relevant layers after each step. This is especially useful because the images change size when the data are resliced and the surface is extracted.