Make compatible with cropped data and display laser position
The code initial had some hard coded values for numbers that work with TOMCAT data, however these caused problem with a smaller dataset cropped in time and space. Furthermore, it adds a blue line on top of the window resliced image to indicate the estimated laser position. this makes it easier to adjust the window offset and size.