Add slider for kernel sizes in laser pos and speed detection
The detection of the laser position and speed has two main parameters, the kernel size for the median filter applied to the image and the kernel size of the median filter applied to the maxima positions. They can now be changed with sliders in the gui. The user can choose to overwrite old results when the function is run multiple times and the function is automatically rerun when the sliders are moved.