Fix widget test by seting the input comboboxes for each step
The test was failing because it relied on the fact that all output layers are named correctly and that the next step uses the correct input layer just based on the name. This does not work anymore since each step now as a drop down menu to select the input layers.