Remove splitting
Decided burden of preparing the data on the end user. Fiji can be used to crop the image to make processing faster, split it to only include one laser pass and to bin it if denoising is required. However, I will keep the h5 reader as part of this plugin to make the conversion from h5 to tif easier.